Fixed camera moving with animation | Fixed broken axe textures and added latest axe model
I fixed the camera moving with the animation by disabling head-bobbling - some code that the FPSController script depends on resets the camera position, but I didn't have time to look deeper into the dependencies to find where. There are better fixes we can look into involving disabling the aforementioned code that resets the position, once we can find it - or potentially rewriting the FPSController to fit our needs more specifically and or making a custom camera from scratch either of these options seem good in the long run as long as we can return head bobbing imo.

I fixed the axe textures/normmaps and added the newest model to the players hand, however you can't see much of it due to the pose of the player by mixamorig- I'd like to get custom animations up as soon as possible - this is @illusionman1212 's responsibility